+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - Replace theme and icon on LocationMap - luc
 - Removes "View on Map" text from More Shows section in Show view - ashley
 - Adds more data marshalling for city view, sets up city scaffold - ash
+- Readds the overlay to fair banner image - kieran
 
 ### 1.8.0